requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-config-url==0.1.1
Django==5.2.7
django-allauth==65.13.0
django-environ==0.12.0
django-seed==0.3.1
djangorestframework==3.16.1
Faker==37.11.0
idna==3.11
oauthlib==3.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyMySQL==1.1.2
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
